Item 5 in issue # 11 - Max execution time.
This line is for 5000 + data points. For less than 5000, this line can
be omitted. It is possible to move this into a separate step run as
tribal job - I think using sheer computing power of the server would be
way faster and reliable plus we get realtime progress, but as a feature
request (given time) since I feel that such move will greatly affect
the code base of upload module.

Agreed on May 15, 2017 1:05 PM:
- Validation as tribal job as feature request.
- Reduce the max execution time to half (10 mins) and ensure they are
in the if statements to prevent spillage to other modules.

A thought:
This step can be subjected to numerous repetition, and steps to
Register a job, wait for job queue and execute in tribal job each time
might cause unnecessary wait time to user and might not give a
relatively quick response as what we currently have.