Item 6 in issue # 11 - hook_file_validate() called for more that just the module.
It’s a shame, issue has increased the execution time on KP :(

I have replaced implementation of hook_file_validate() and replaced it
with a custom field upload validators (#upload_validators) in the form
definition in hook_forms. I am hoping this will localize this function
to just this form and to this module only.

Please Review:
I am unsure of passing $form_state to an upload_validators function
callback but what I want is the file, the field will contain after drag
and drop. $form_state appears to contain the file object in this case.